requirements.txt
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
dnspython==1.16.0
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
pip==21.2.4
Pillow==9.0.0
pymongo==3.12.0
python-dateutil==2.8.2
s3transfer==0.5.0
setuptools==58.0.4
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
wheel==0.37.0
wincertstore==0.2
gunicorn
